inventory wip
