+++ src/lang/en.json
@@ -300,6 +300,30 @@
   "TB2.UserInteractionDeleteItemTitle": "Delete Item",
   "TB2.Description": "Description",
   "TB2.ItemName": "Name",
+  "TB2.HeadSlotHeader": "Head",
+  "TB2.CustomSlotHeader": "PLACEHOLDER",
+  "TB2.PackSlotHeader": "Pack",
+  "TB2.HeldSlotHeader": "Held",
+  "TB2.PocketSlotHeader": "Pocket",
+  "TB2.BeltSlotHeader": "Belt",
+  "TB2.TorsoSlotHeader": "Torso",
+  "TB2.FeetSlotHeader": "Feet",
+  "TB2.HandSlotHeader": "Hands",
+  "TB2.NeckSlotHeader": "Neck",
+  "TB2.XOfYLabel": "{0} / {1}",
+  "TB2.": "PLACEHOLDER",
+  "TB2.": "PLACEHOLDER",
+  "TB2.": "PLACEHOLDER",
+  "TB2.": "PLACEHOLDER",
+  "TB2.": "PLACEHOLDER",
+  "TB2.": "PLACEHOLDER",
+  "TB2.": "PLACEHOLDER",
+  "TB2.": "PLACEHOLDER",
+  "TB2.": "PLACEHOLDER",
+  "TB2.": "PLACEHOLDER",
+  "TB2.": "PLACEHOLDER",
+  "TB2.": "PLACEHOLDER",
+  "TB2.": "PLACEHOLDER",
   "TB2.": "PLACEHOLDER",
   "TB2.": "PLACEHOLDER",
   "TB2.": "PLACEHOLDER",
