+++ src/module/actor/TBActorDataProperties.ts
@@ -36,6 +36,24 @@ interface HasMight {
   might: number;
   precedence: number;
 }
+// the intent here is to have anything that has an inventory (i.e. player character) can define
+// how many slots and of what type they contain. or, actor item slots represent _capacity_, and
+// item inventory slots represent volume against that capacity.
+export interface HasInventorySlots {
+  containerSlots: {
+    pack: number;
+    held: number;
+    belt: number;
+    torso: number;
+    head: number;
+    hand: number;
+    feet: number;
+    neck: number;
+    arm: number;
+    // as pockets are effectively infinite, this represents the capability to have item slots at all
+    pocket: boolean;
+  };
+}
 
 type TBActorDataPropertiesDataAttributes = {
   [Key in keyof typeof TB.i18n.attributes]: ModifiableDataBaseTotal<number>;
@@ -61,9 +79,11 @@ export function isTest(value: string): value is SkillTest {
 interface TBCreatureDataPropertiesData
   extends TBActorDataPropertiesDataBase,
     HasMight,
+    HasInventorySlots,
     CharacterDataSource {}
 
 interface TBCharacterDataPropertiesData
   extends TBActorDataPropertiesDataBase,
     HasMight,
+    HasInventorySlots,
     NpcDataSourceData {}
